pyproject.toml
[tool.poetry]
name = "derakht-backend"
version = "0.1.0"
description = ""
authors = ["Arian Hedayati Far <arian.hedayatifar@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.1.2"
psycopg2-binary = "^2.9.10"
djangorestframework = "^3.15.2"
wagtail = "^6.2.2"
wagtail-modeladmin = "^2.0.0"
jdatetime = "^5.0.0"
django-storages = "^1.14.4"
boto3 = "^1.35.54"
djangorestframework-simplejwt = "^5.3.1"
django-phonenumber-field = "^8.0.0"
pillow = "^11.0.0"
django-phonenumbers = "^1.0.1"
django-cors-headers = "^4.6.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
